backend/requirements.txt
fastapi==0.136.1
uvicorn==0.35.0
sqlalchemy
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
psycopg2-binary
python-dotenv
pandas
sentence-transformers
gunicorn